+++ CHANGELOG.md
@@ -7,3 +7,4 @@ master
 * Add monolog processor to encrypt logs
 * Add optional sonata admin to decrypt logs
 * Add processor to decrypt secrets at runtime
+* Add optional sonata admin to encrypt content